Refactor JS code
