+++ js/src/common.js
@@ -11,15 +11,6 @@ export const INS = {
   SET_SLOT: 0x12,
 };
 
-export const PAYLOAD_TYPE = {
-  INIT: 0x00,
-  ADD: 0x01,
-  LAST: 0x02,
-  TEMPLATE: 0x03,
-  MERKLE_TREE: 0x04,
-  MERKLE_TREE_LAST: 0x05,
-};
-
 export const P1_VALUES = {
   ONLY_RETRIEVE: 0x00,
   SHOW_ADDRESS_IN_DEVICE: 0x01,
