+++ js/src/index.js
@@ -15,21 +15,9 @@
  *  limitations under the License.
  ******************************************************************************* */
 
-import {serializePathv1, PATH_SERIALIZATION_VERSION, printBIP44Path, signSendChunkv1, validateCryptoOptions} from "./helperV1";
-import {
-  CHUNK_SIZE,
-  CLA,
-  ERROR_CODE,
-  errorCodeToString,
-  getVersion,
-  INS,
-  P1_VALUES,
-  PKLEN,
-  processErrorResponse,
-  compareVersion,
-} from "./common";
-import { signSendChunkv2 } from "./helperV2";
-import { merkleIndex, merkleTree } from "./txMerkleTree"
+import { serializePath, validateCryptoOptions, printBIP44Path} from "./serializePath";
+import { signGetChunks, signIsLastAPDU } from "./signStoredTransaction";
+import { CLA, ERROR_CODE, errorCodeToString, getVersion, INS, P1_VALUES, PKLEN, processErrorResponse } from "./common";
 
 function processGetAddrResponse(response) {
   let partialResponse = response;
@@ -74,58 +62,7 @@ export default class FlowApp {
     }
     this.transport = transport;
   }
-
-  static preparePathAndTxChunks(serializedPathBuffer, message) {
-    const chunks = [];
-
-    // First chunk (only path)
-    chunks.push(serializedPathBuffer);
-
-    const messageBuffer = Buffer.from(message);
-
-    const buffer = Buffer.concat([messageBuffer]);
-    for (let i = 0; i < buffer.length; i += CHUNK_SIZE) {
-      let end = i + CHUNK_SIZE;
-      if (i > buffer.length) {
-        end = buffer.length;
-      }
-      chunks.push(buffer.slice(i, end));
-    }
-
-    return chunks;
-  }
-
-  signGetChunksv1(path, options, pathSerializationVersion, message) {
-    const serializedPath = serializePathv1(path, pathSerializationVersion, options);
-    return FlowApp.preparePathAndTxChunks(serializedPath, message);
-  }
-
-  signGetChunksv2(path, options, pathSerializationVersion, message, scriptHash) {
-    const serializedPath = serializePathv1(path, pathSerializationVersion, options);
-    const basicChunks =  FlowApp.preparePathAndTxChunks(serializedPath, message);
-
-    //other chunks
-    const merkleI = merkleIndex[scriptHash.slice(0, 16)]
-    if (merkleI === undefined) {
-        return undefined
-    }
-    const metadata = merkleTree.children[merkleI[0]].children[merkleI[1]].children[merkleI[2]].children[merkleI[3]].children[0]
-    const merkleTreeLevel1 = merkleTree.children[merkleI[0]].children[merkleI[1]].children[merkleI[2]].children.map((ch) => ch.hash).join('')
-    const merkleTreeLevel2 = merkleTree.children[merkleI[0]].children[merkleI[1]].children.map((ch) => ch.hash).join('')
-    const merkleTreeLevel3 = merkleTree.children[merkleI[0]].children.map((ch) => ch.hash).join('')
-    const merkleTreeLevel4 = merkleTree.children.map((ch) => ch.hash).join('')
-
-    return [
-      ...basicChunks, 
-      Buffer.from(metadata, "hex"), 
-      Buffer.from(merkleTreeLevel1, "hex"),
-      Buffer.from(merkleTreeLevel2, "hex"),
-      Buffer.from(merkleTreeLevel3, "hex"),
-      Buffer.from(merkleTreeLevel4, "hex"),
-    ]
-  }
   
-
   async getVersion() {
     return getVersion(this.transport)
       .then((response) => {
@@ -191,10 +128,8 @@ export default class FlowApp {
 
   async getAddressAndPubKey(path, cryptoOptions) {
     validateCryptoOptions(cryptoOptions);
-
     const getVersionResponse = await this.getVersion();
-    const pathSerializationVersion = this.getPathSetializationVersion(getVersionResponse);
-    const serializedPath = serializePathv1(path, pathSerializationVersion, cryptoOptions);
+    const serializedPath = serializePath(path, getVersionResponse, cryptoOptions);
 
     return this.transport
       .send(CLA, INS.GET_PUBKEY, P1_VALUES.ONLY_RETRIEVE, 0, serializedPath, [0x9000])
@@ -203,95 +138,91 @@ export default class FlowApp {
 
   async showAddressAndPubKey(path, cryptoOptions) {
     validateCryptoOptions(cryptoOptions);
-
     const getVersionResponse = await this.getVersion();
-    const pathSerializationVersion = this.getPathSetializationVersion(getVersionResponse);
-
-    const serializedPath = serializePathv1(path, pathSerializationVersion, cryptoOptions);
+    const serializedPath = serializePath(path, getVersionResponse, cryptoOptions);
 
     return this.transport
       .send(CLA, INS.GET_PUBKEY, P1_VALUES.SHOW_ADDRESS_IN_DEVICE, 0, serializedPath, [0x9000])
       .then(processGetAddrResponse, processErrorResponse);
   }
 
-  async signSendChunk(chunkIdx, chunkNum, chunk) {
-    return signSendChunkv1(this, chunkIdx, chunkNum, chunk);
-  }
-
-  async signSendChunkv2(chunkIdx, chunkNum, chunk) {
-    return signSendChunkv2(this, chunkIdx, chunkNum, chunk);
-  }
-
   async sign(path, message, cryptoOptions, scriptHash) {
     validateCryptoOptions(cryptoOptions);
-
     const getVersionResponse = await this.getVersion();
-    const pathSerializationVersion = this.getPathSetializationVersion(getVersionResponse);
+    const chunks = signGetChunks(path, cryptoOptions, getVersionResponse, message, scriptHash)
 
-    // APDU call sequence without Merkle trees
-    if (compareVersion(getVersionResponse, 0, 10, 3) <= 0) {
-      const chunks = this.signGetChunksv1(path, cryptoOptions, pathSerializationVersion, message)
-      return this.signSendChunk(1, chunks.length, chunks[0]).then(async (response) => {
-        let result = {
-          returnCode: response.returnCode,
-          errorMessage: response.errorMessage,
-          signatureCompact: null,
-          signatureDER: null,
-        };
+    if (typeof chunks === "string") {
+      return {
+        signatureCompact: null,
+        signatureDER: null,
+        returnCode: 0xffff,
+        errorMessage: chunks,
+      }
+    }
 
-        for (let i = 1; i < chunks.length; i += 1) {
-          // eslint-disable-next-line no-await-in-loop
-          result = await this.signSendChunk(1 + i, chunks.length, chunks[i]);
-          if (result.returnCode !== ERROR_CODE.NoError) {
-            break;
+    let result = {
+      signatureCompact: null,
+      signatureDER: null,
+      returnCode: 0xffff,
+      errorMessage: "Result not initialized",
+    }
+
+    for (let i = 0; i < chunks.length; i += 1) {
+      const payloadType = chunks[i].type
+      const chunk = chunks[i].buffer
+
+      // eslint-disable-next-line no-await-in-loop
+      result = await this.transport
+        .send(CLA, INS.SIGN, payloadType, 0, chunk, [0x9000, 0x6984, 0x6a80])
+        .then((response) => {
+          const errorCodeData = response.slice(-2);
+          const returnCode = errorCodeData[0] * 256 + errorCodeData[1];
+          const errorMessage = errorCodeToString(returnCode);
+
+          // these error codes contain detailed error description in the response
+          if (returnCode === 0x6a80 || returnCode === 0x6984) {
+            const detailedErrorMessage = `${errorMessage} : ${response.slice(0, response.length - 2).toString("ascii")}`;
+            return {
+              signatureCompact: null,
+              signatureDER: null,
+              returnCode: returnCode,
+              errorMessage: detailedErrorMessage,
+            };
           }
-        }
 
-        return {
-          returnCode: result.returnCode,
-          errorMessage: result.errorMessage,
-          // ///
-          signatureCompact: result.signatureCompact,
-          signatureDER: result.signatureDER,
-        };
-      }, processErrorResponse);
-    }
+          // last APDU, everything OK
+          if (returnCode == 0x9000 && signIsLastAPDU(payloadType)) {
+            const signatureCompact = response.slice(0, 65);
+            const signatureDER = response.slice(65, response.length - 2);
+            return {
+              signatureCompact,
+              signatureDER,
+              returnCode: returnCode,
+              errorMessage: errorMessage,
+            };  
+          }    
 
-    // APDU call sequence with Merkle trees
-    const chunks = this.signGetChunksv2(path, cryptoOptions, pathSerializationVersion, message, scriptHash)
-    if (chunks === undefined) {
-      return {
-        returnCode: 0xffff,  //used for JS errors
-        errorMessage: "Script hash not known",
-        signatureCompact: null,
-        signatureDER: null,
+          // Here return code should be 0x9000 as for error codes other than [0x9000, 0x6984, 0x6a80] send throws
+          return {
+            signatureCompact: null,
+            signatureDER: null,
+            returnCode: returnCode,
+            errorMessage: errorMessage,
+          };
+        }, processErrorResponse);
+        
+      // If any error accours we do not send further APDUs
+      if (result.returnCode !== ERROR_CODE.NoError) {
+        break;
       }
     }
 
-    return this.signSendChunkv2(1, chunks.length, chunks[0]).then(async (response) => {
-      let result = {
-        returnCode: response.returnCode,
-        errorMessage: response.errorMessage,
-        signatureCompact: null,
-        signatureDER: null,
-      };
-
-      for (let i = 1; i < chunks.length; i += 1) {
-        // eslint-disable-next-line no-await-in-loop
-        result = await this.signSendChunkv2(1 + i, chunks.length, chunks[i]); 
-        if (result.returnCode !== ERROR_CODE.NoError) {
-          break;
-        }
-      }
-
-      return {
-        returnCode: result.returnCode,
-        errorMessage: result.errorMessage,
-        // ///
-        signatureCompact: result.signatureCompact,
-        signatureDER: result.signatureDER,
-      };
-    }, processErrorResponse);
+    return {
+      returnCode: result.returnCode,
+      errorMessage: result.errorMessage,
+      signatureCompact: result.signatureCompact,
+      signatureDER: result.signatureDER,
+    };
   }
 
   async slotStatus() {
@@ -348,9 +279,7 @@ export default class FlowApp {
     const serializedAccount = Buffer.from(account, "hex");
   
     const getVersionResponse = await this.getVersion();
-    const pathSerializationVersion = this.getPathSetializationVersion(getVersionResponse);
-
-    const serializedPath = serializePathv1(path, pathSerializationVersion, cryptoOptions);
+    const serializedPath = serializePath(path, getVersionResponse, cryptoOptions);
 
     if (serializedAccount.length !== 8) {
       return {
